NULL device redirection
Since we work with VistA a lot, and it tends to write to the screen,
open the null device so all output goes to it.

Downstream code should not try to close the null device. It can try
re-opening it, but may not succeed. That's okay to not succeed in
opening it.

The device is %WNULL.

I am trying to think of a better way of doing this where VistA can still
use NULL devices without causing any problems.