Fix regression in 58a9e222df8d11a9ddac2fd161711ddc41ac8296
We checked for failure of getting a web page using $D(HTTPERR). That was
not correct. HTTPERR is always defined and equal to zero in non-error
cases. The correct way is to use an if statement on it.

As a result, we couldn't get index.html pages anymore.

And it turns out there was no explicit test for index.html pages. Now
added.