[#59] Zipping .woff2 returns incorrect data to the web browser/client
It turns out that the code to write out a binary file after it was read
was missing `set $x=0`, otherwise GT.M/YottaDB by default will add a
LF to the file at the end. This caused the following steps of gzipping
the binary file to effectively fail.

No tests, as I can't think of an easy test to test this within the
framework used in the Web Server.

Version bumped to 1.1.4.