[#55] Non UTF-8 byte sequences in body causes crashes
I have not been able to reproduce any crashes in my testing; however,
the read loop in %webreq (see the changes for the exact line) never
finished because it relied on $LENGTH to figure out how many bytes were
read, and in GT.M/YottaDB UTF-8 mode, $LENGTH does not get the number of
bytes, but rather, the number of code points.

The fix is for GT.M, use $ZLENGTH.

The other uses of $LENGTH are okay as they either read the first
character (to determine if it is a "^"), or they are used to determine
if there is data.

Test `/test/utf8/post` added to test Posting UTF8 data and getting a
result back. It's in `tpostutf8^%webtest`.